refactor: 장바구니 상품 수량 변경 콜백 수정

+++ app/src/main/java/woowacourse/shopping/data/datasource/CartItemDataSource.kt
@@ -18,13 +18,13 @@ interface CartItemDataSource {
 
     fun removeCartItem(
         cartId: Long,
-        callback: (Long) -> Unit,
+        callback: () -> Unit,
     )
 
     fun updateCartItem(
         cartId: Long,
         quantity: Quantity,
-        onResult: (Long) -> Unit,
+        onResult: () -> Unit,
     )
 
     fun fetchCartItemsCount(onResult: (Quantity?) -> Unit)
